OpenClaw-MemBench/requirements.txt
pyyaml>=6.0
rich>=13.7.0
python-dotenv>=1.0.1
requests>=2.32.3
